refactor: import React
