+++ src/app/[locale]/(public)/@modal/(.)photo-details/[id]/page.tsx
@@ -1,4 +1,5 @@
 import type { Locale } from 'next-intl'
+import { setRequestLocale } from 'next-intl/server'
 import { getPhotoById } from '@/shared/api'
 import { PhotoDetailsModal } from '@/widgets/photo-details-modal'
 
@@ -8,9 +9,11 @@ type Props = {
 }
 
 export default async function Page(props: Props) {
-	const { id } = await props.params
+	const { id, locale } = await props.params
 
 	const { data, error } = await getPhotoById({ id })
+
+	setRequestLocale(locale)
 
 	if (error) {
 		console.error('Error fetching photo details:', error)
