+++ src/app/[locale]/error.tsx
@@ -18,10 +18,10 @@ export default function Error({ error, reset }: Props) {
 	return (
 		<div>
 			{t.rich('description', {
-				p: chunks => <p className='mt-4'>{chunks}</p>,
+				p: chunks => <p>{chunks}</p>,
 				retry: chunks => (
 					<button
-						className='text-white underline underline-offset-2'
+						className='underline underline-offset-2'
 						onClick={reset}
 						type='button'
 					>
