+++ src/app/layout.tsx
@@ -1,5 +1,6 @@
+import type { ReactNode } from 'react'
 import './styles/index.css'
 
-export default async function RootLayout({ children }: Readonly<{ children: React.ReactNode }>) {
+export default async function RootLayout({ children }: Readonly<{ children: ReactNode }>) {
 	return children
 }
